requirements.txt
# Twitter/X scraping
yt-dlp>=2024.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Video processing
moviepy>=1.0.3
opencv-python>=4.9.0
numpy>=1.24.0

# Audio analysis (for detecting impactful moments)
librosa>=0.10.0
soundfile>=0.12.0

# AI for text reformulation (optional - can use free alternatives)
openai>=1.0.0
# Alternative: use local models with transformers
# transformers>=4.36.0
# torch>=2.0.0

# TikTok upload
playwright>=1.40.0
TikTokApi>=6.0.0

# Scheduling and monitoring
schedule>=1.2.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Image/video analysis
Pillow>=10.0.0
